% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_na_columns}
\alias{add_na_columns}
\title{Add NA columns to row tibble}
\usage{
add_na_columns(rows_tbl, col_metadata)
}
\arguments{
\item{rows_tbl}{Tibble with group and row_label columns}

\item{col_metadata}{tibble with col_id column}
}
\value{
Tibble with rows, row labels, and NA columns for each col_id
}
\description{
Add NA columns to row tibble
}
\keyword{internal}
