% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_columns.R
\name{construct_columns}
\alias{construct_columns}
\title{Construct column metadata from col_spec}
\usage{
construct_columns(data, col_spec)
}
\arguments{
\item{data}{Data frame with labelled variables}

\item{col_spec}{List of column specifications}
}
\value{
Tibble with columns: col_block, col_level_1, col_level_2, col_id
}
\description{
Construct column metadata from col_spec
}
\keyword{internal}
