% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_rows.R
\name{construct_rows}
\alias{construct_rows}
\title{Construct row structure from by_list}
\usage{
construct_rows(data, by_list)
}
\arguments{
\item{data}{Data frame with labelled variables}

\item{by_list}{Named list defining row hierarchy}
}
\value{
Tibble with columns: group, row_label
}
\description{
Construct row structure from by_list
}
\keyword{internal}
