% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_blocks_in_order}
\alias{get_blocks_in_order}
\title{Get unique blocks in order}
\usage{
get_blocks_in_order(col_metadata)
}
\arguments{
\item{col_metadata}{Tibble with column metadata}
}
\value{
Character vector of unique col_block values in order of appearance
}
\description{
Get unique blocks in order
}
\keyword{internal}
