% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_col_block_structure}
\alias{get_col_block_structure}
\title{Get column block structure}
\usage{
get_col_block_structure(col_metadata)
}
\arguments{
\item{col_metadata}{Tibble with column metadata}
}
\value{
List with structure: list(blocks = list(block_name = list(label_1 = col_id, ...)), ...)
}
\description{
Returns information about column blocks (spanners) and their columns.
}
\keyword{internal}
