% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_col_ids_for_block}
\alias{get_col_ids_for_block}
\title{Get column IDs for a block}
\usage{
get_col_ids_for_block(col_metadata, block_name)
}
\arguments{
\item{col_metadata}{Tibble with column metadata}

\item{block_name}{Character string for block name}
}
\value{
Character vector of col_ids
}
\description{
Get column IDs for a block
}
\keyword{internal}
