% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_col_ids_for_level_1}
\alias{get_col_ids_for_level_1}
\title{Get column IDs for a specific col_level_1 within a block}
\usage{
get_col_ids_for_level_1(col_metadata, block_name, level_1_label)
}
\arguments{
\item{col_metadata}{Tibble with column metadata}

\item{block_name}{Character string for block name}

\item{level_1_label}{Character string for col_level_1 value}
}
\value{
Character vector of col_ids
}
\description{
Get column IDs for a specific col_level_1 within a block
}
\keyword{internal}
