% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_col_labels}
\alias{get_col_labels}
\title{Get column labels for a set of col_ids}
\usage{
get_col_labels(col_metadata, col_ids)
}
\arguments{
\item{col_metadata}{Tibble with column metadata}

\item{col_ids}{Character vector of col_ids}
}
\value{
Named character vector with col_id = label
}
\description{
Get column labels for a set of col_ids
}
\keyword{internal}
