% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_unique_level_1_for_block}
\alias{get_unique_level_1_for_block}
\title{Get unique col_level_1 values for a block}
\usage{
get_unique_level_1_for_block(col_metadata, block_name)
}
\arguments{
\item{col_metadata}{Tibble with column metadata}

\item{block_name}{Character string for block name}
}
\value{
Character vector of unique non-NA col_level_1 values in order of appearance
}
\description{
Get unique col_level_1 values for a block
}
\keyword{internal}
