% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_col_level_2}
\alias{has_col_level_2}
\title{Check if a block has subgroups (col_level_2)}
\usage{
has_col_level_2(col_metadata, block_name)
}
\arguments{
\item{col_metadata}{Tibble with column metadata}

\item{block_name}{Character string for block name}
}
\value{
Logical; TRUE if the block has any non-NA col_level_2 values
}
\description{
Check if a block has subgroups (col_level_2)
}
\keyword{internal}
