% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_columns.R
\name{make_valid_names}
\alias{make_valid_names}
\title{Convert character vector to valid R names}
\usage{
make_valid_names(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Character vector of valid R names
}
\description{
Convert character vector to valid R names
}
\keyword{internal}
