% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate inputs to make_table_shell}
\usage{
validate_inputs(data, by_list, col_spec)
}
\arguments{
\item{data}{Data frame to validate}

\item{by_list}{Named list for row structure}

\item{col_spec}{List of column specifications}
}
\description{
Validate inputs to make_table_shell
}
\keyword{internal}
